schemeedit: Move clipboard helpers to utils
