+++ orangecanvas/gui/utils.py
@@ -7,13 +7,14 @@ import traceback
 import ctypes
 
 from contextlib import contextmanager
+from typing import Optional
 
 from AnyQt.QtWidgets import (
     QWidget, QMessageBox, QStyleOption, QStyle, QTextEdit, QScrollBar
 )
 from AnyQt.QtGui import (
     QGradient, QLinearGradient, QRadialGradient, QBrush, QPainter,
-    QPaintEvent, QColor, QPixmap, QPixmapCache, QTextOption
+    QPaintEvent, QColor, QPixmap, QPixmapCache, QTextOption, QGuiApplication
 )
 from AnyQt.QtCore import Qt, QPointF, QPoint, QRect, QRectF, Signal, QEvent
 
@@ -540,3 +541,29 @@ def innerShadowPixmap(color, size, pos, length=5):
     QPixmapCache.insert(key, finalShadow)
 
     return finalShadow
+
+
+def clipboard_has_format(mimetype):
+    # type: (str) -> bool
+    """Does the system clipboard contain data for mimetype?"""
+    cb = QGuiApplication.clipboard()
+    if cb is None:
+        return False
+    mime = cb.mimeData()
+    if mime is None:
+        return False
+    return mime.hasFormat(mimetype)
+
+
+def clipboard_data(mimetype: str) -> Optional[bytes]:
+    """Return the binary data of the system clipboard for mimetype."""
+    cb = QGuiApplication.clipboard()
+    if cb is None:
+        return None
+    mime = cb.mimeData()
+    if mime is None:
+        return None
+    if mime.hasFormat(mimetype):
+        return bytes(mime.data(mimetype))
+    else:
+        return None
